pyproject.toml
[tool.poetry]
name = "meadowdata"
version = "0.1.0"
description = "A platform for data scientists, including a database and a job scheduler"
authors = ["Richard Lee <hrichardlee@gmail.com>"]
homepage = "https://github.com/meadowdata/meadowdata"
repository = "https://github.com/meadowdata/meadowdata"
keywords = []
packages = [
    { include = "meadowflow", from = "src" },
    { include = "meadowdb", from = "src" },
    { include = "meadowrun", from = "src" },
]

[tool.poetry.scripts]
meadowrun = 'meadowrun.server_main:command_line_main'
meadowflow = 'meadowflow.server.server_main:command_line_main'

[tool.poetry.dependencies]
python = "^3.9"
duckdb = "^0.2.8"
pyarrow = "^5.0.0"
fastparquet = "^0.7.1"
pandas = "^1.3.1"
grpcio = "^1.39.0"
protobuf = "^3.18.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pylint = "^2.8.2"
mypy = "^0.812"
black = {version = "^21.6b0", allow-prereleases = true}
grpcio-tools = "^1.39.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
